refactor(encryption): deprecate existing encryption system
done in favor of introducing a new system
app will still use this system for now
tests also
removed
