+++ source/services/encryption/index.js
@@ -1,5 +1,5 @@
-export { encryptFormAnswers } from "./EncryptionService";
-export { decryptFormAnswers } from "./EncryptionService";
-export { encryptWithAesKey } from "./EncryptionService";
-export { decryptWithAesKey } from "./EncryptionService";
-export { setupSymmetricKey } from "./EncryptionService";
+export { encryptFormAnswers } from "./DEPRECATED/EncryptionService";
+export { decryptFormAnswers } from "./DEPRECATED/EncryptionService";
+export { encryptWithAesKey } from "./DEPRECATED/EncryptionService";
+export { decryptWithAesKey } from "./DEPRECATED/EncryptionService";
+export { setupSymmetricKey } from "./DEPRECATED/EncryptionService";
